move utility to seed test db to libs/model
