+++ libs/model/src/index.ts
@@ -9,6 +9,9 @@ export * as User from './user';
 export * from './services/commonProtocol';
 export * from './services/tokenBalanceCache';
 
+// Test Service
+export * as tester from './test';
+
 // Internals - Should not be exported once we finish the migrations to models
 export * from './database';
 export * from './models';
