+++ packages/commonwealth/test/integration/api/chainNodes.spec.ts
@@ -1,13 +1,12 @@
 import { BalanceType } from '@hicommonwealth/core';
-import { models, UserInstance } from '@hicommonwealth/model';
+import { models, tester, UserInstance } from '@hicommonwealth/model';
 import { assert, expect } from 'chai';
-import { resetDatabase } from 'server-test';
 import { ServerCommunitiesController } from '../../../server/controllers/server_communities_controller';
 import { buildUser } from '../../unit/unitHelpers';
 
 describe('ChainNode Tests', () => {
   beforeEach(async () => {
-    await resetDatabase();
+    await tester.seedDb();
   });
 
   it('Creates new ChainNode when', async () => {
