+++ packages/commonwealth/test/integration/api/index.spec.ts
@@ -1,14 +1,14 @@
 /* eslint-disable no-unused-expressions */
 import { ChainBase } from '@hicommonwealth/core';
+import { tester } from '@hicommonwealth/model';
 import { personalSign } from '@metamask/eth-sig-util';
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-
 import wallet from 'ethereumjs-wallet';
 import { ethers } from 'ethers';
 import { bech32ToHex } from 'shared/utils';
 import * as siwe from 'siwe';
-import app, { resetDatabase } from '../../../server-test';
+import app from '../../../server-test';
 import {
   TEST_BLOCK_INFO_BLOCKHASH,
   TEST_BLOCK_INFO_STRING,
@@ -22,7 +22,7 @@ const { expect } = chai;
 
 describe('API Tests', () => {
   before('reset database', async () => {
-    await resetDatabase();
+    await tester.seedDb();
   });
 
   describe('address tests', () => {
