+++ packages/commonwealth/test/integration/api/updateChain.spec.ts
@@ -4,12 +4,12 @@ import {
   TokenBalanceCache,
   UserInstance,
   models,
+  tester,
 } from '@hicommonwealth/model';
 import { assert } from 'chai';
 import { ServerCommunitiesController } from '../../../server/controllers/server_communities_controller';
 import { Errors } from '../../../server/controllers/server_communities_methods/update_community';
 import { buildUser } from '../../unit/unitHelpers';
-import { resetDatabase } from '../../util/resetDatabase';
 
 const baseRequest: CommunityAttributes = {
   id: 'ethereum',
@@ -26,7 +26,7 @@ const baseRequest: CommunityAttributes = {
 
 describe('UpdateChain Tests', () => {
   before(async () => {
-    await resetDatabase();
+    await tester.seedDb();
   });
 
   it('Correctly updates chain', async () => {
