+++ packages/commonwealth/test/integration/databaseCleaner.spec.ts
@@ -1,10 +1,9 @@
 import { NotificationCategories } from '@hicommonwealth/core';
-import { models } from '@hicommonwealth/model';
+import { models, tester } from '@hicommonwealth/model';
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import { Sequelize } from 'sequelize';
 import sinon from 'sinon';
-import { resetDatabase } from '../../server-test';
 import DatabaseCleaner from '../../server/util/databaseCleaner';
 
 chai.use(chaiHttp);
@@ -12,7 +11,7 @@ const { expect } = chai;
 
 describe('DatabaseCleaner Tests', () => {
   before('Reset database', async () => {
-    await resetDatabase();
+    await tester.seedDb();
   });
 
   describe('Tests when the cleaner runs', () => {
